requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
decorator==5.1.1
idna==3.3
pyTelegramBotAPI==4.5.0
python-dotenv==0.20.0
requests==2.27.1
urllib3==1.26.9
websockets==10.3
